+++ bezier/src/cubic_bezier.rs
@@ -1,0 +1,141 @@
+use euclid::rect;
+use {Point, Rect};
+use up_to_two::UpToTwo;
+use flatten_cubic::{flatten_cubic_bezier, find_cubic_bezier_inflection_points};
+pub use flatten_cubic::CubicFlatteningIter;
+pub use cubic_to_quadratic::cubic_to_quadratic;
+
+/// A 2d curve segment defined by four points: the beginning of the segment, two control
+/// points and the end of the segment.
+///
+/// The curve is defined by equation:²
+/// ```∀ t ∈ [0..1],  P(t) = (1 - t)³ * from + 3 * (1 - t)² * t * ctrl1 + 3 * t² * (1 - t) * ctrl2 + t³ * to```
+#[derive(Copy, Clone, Debug)]
+pub struct CubicBezierSegment {
+    pub from: Point,
+    pub ctrl1: Point,
+    pub ctrl2: Point,
+    pub to: Point,
+}
+
+impl CubicBezierSegment {
+    /// Sample the curve at t (expecting t between 0 and 1).
+    pub fn sample(&self, t: f32) -> Point {
+        let t2 = t * t;
+        let t3 = t2 * t;
+        let one_t = 1.0 - t;
+        let one_t2 = one_t * one_t;
+        let one_t3 = one_t2 * one_t;
+        return self.from * one_t3 +
+            self.ctrl1.to_vector() * 3.0 * one_t2 * t +
+            self.ctrl2.to_vector() * 3.0 * one_t * t2 +
+            self.to.to_vector() * t3;
+    }
+
+    /// Split this curve into two sub-curves.
+    pub fn split(&self, t: f32) -> (CubicBezierSegment, CubicBezierSegment) {
+        let ctrl1a = self.from + (self.ctrl1 - self.from) * t;
+        let ctrl2a = self.ctrl1 + (self.ctrl2 - self.ctrl1) * t;
+        let ctrl1aa = ctrl1a + (ctrl2a - ctrl1a) * t;
+        let ctrl3a = self.ctrl2 + (self.to - self.ctrl2) * t;
+        let ctrl2aa = ctrl2a + (ctrl3a - ctrl2a) * t;
+        let ctrl1aaa = ctrl1aa + (ctrl2aa - ctrl1aa) * t;
+        let to = self.to;
+
+        return (CubicBezierSegment {
+            from: self.from,
+            ctrl1: ctrl1a,
+            ctrl2: ctrl1aa,
+            to: ctrl1aaa,
+        },
+        CubicBezierSegment {
+            from: ctrl1aaa,
+            ctrl1: ctrl2aa,
+            ctrl2: ctrl3a,
+            to: to,
+        });
+    }
+
+    /// Return the curve before the split point.
+    pub fn before_split(&self, t: f32) -> CubicBezierSegment {
+        let ctrl1a = self.from + (self.ctrl1 - self.from) * t;
+        let ctrl2a = self.ctrl1 + (self.ctrl2 - self.ctrl1) * t;
+        let ctrl1aa = ctrl1a + (ctrl2a - ctrl1a) * t;
+        let ctrl3a = self.ctrl2 + (self.to - self.ctrl2) * t;
+        let ctrl2aa = ctrl2a + (ctrl3a - ctrl2a) * t;
+        let ctrl1aaa = ctrl1aa + (ctrl2aa - ctrl1aa) * t;
+        return CubicBezierSegment {
+            from: self.from,
+            ctrl1: ctrl1a,
+            ctrl2: ctrl1aa,
+            to: ctrl1aaa,
+        };
+    }
+
+    /// Return the curve after the split point.
+    pub fn after_split(&self, t: f32) -> CubicBezierSegment {
+        let ctrl1a = self.from + (self.ctrl1 - self.from) * t;
+        let ctrl2a = self.ctrl1 + (self.ctrl2 - self.ctrl1) * t;
+        let ctrl1aa = ctrl1a + (ctrl2a - ctrl1a) * t;
+        let ctrl3a = self.ctrl2 + (self.to - self.ctrl2) * t;
+        let ctrl2aa = ctrl2a + (ctrl3a - ctrl2a) * t;
+        return CubicBezierSegment {
+            from: ctrl1aa + (ctrl2aa - ctrl1aa) * t,
+            ctrl1: ctrl2a + (ctrl3a - ctrl2a) * t,
+            ctrl2: ctrl3a,
+            to: self.to,
+        };
+    }
+
+    /// Returns the flattened representation of the curve as an iterator, starting *after* the
+    /// current point.
+    pub fn flattening_iter(&self, tolerance: f32) -> CubicFlatteningIter {
+        CubicFlatteningIter::new(*self, tolerance)
+    }
+
+    /// Iterates through the curve invoking a callback at each point.
+    pub fn flattened_for_each<F: FnMut(Point)>(&self, tolerance: f32, call_back: &mut F) {
+        flatten_cubic_bezier(*self, tolerance, call_back);
+    }
+
+    /// Compute the length of the segment using a flattened approximation.
+    pub fn compute_length(&self, tolerance: f32) -> f32 {
+        let mut start = self.from;
+        let mut len = 0.0;
+        self.flattened_for_each(tolerance, &mut|p| {
+            len += (p - start).length();
+            start = p;
+        });
+        return len;
+    }
+
+    pub fn find_inflection_points(&self) -> UpToTwo<f32> {
+        find_cubic_bezier_inflection_points(self)
+    }
+
+    pub fn bounding_rect(&self) -> Rect {
+        let min_x = self.from.x.min(self.ctrl1.x).min(self.ctrl2.x).min(self.to.x);
+        let max_x = self.from.x.max(self.ctrl1.x).max(self.ctrl2.x).max(self.to.x);
+        let min_y = self.from.y.min(self.ctrl1.y).min(self.ctrl2.y).min(self.to.y);
+        let max_y = self.from.y.max(self.ctrl1.y).max(self.ctrl2.y).max(self.to.y);
+
+        return rect(min_x, min_y, max_x - min_x, max_y - min_y);
+    }
+}
+
+#[test]
+fn bounding_rect_for_cubic_bezier_segment() {
+    let a = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(0.5, 1.0),
+        ctrl2: Point::new(1.5, -1.0),
+        to: Point::new(2.0, 0.0),
+    };
+
+    let expected_bounding_rect = rect(0.0, -1.0, 2.0, 2.0);
+
+    let actual_bounding_rect = a.bounding_rect();
+
+    assert!(expected_bounding_rect == actual_bounding_rect)
+}
+
